New LadderPacket
