+++ KServer/bin/.gitignore
@@ -1,1 +1,8 @@
+/Animations/
 /Kangaroo/
+/Packets/
+/Server/
+/Utils/
+/anims/
+/enums/
+/sprites/
