+++ KServer/src/Packets/HomePacket.java
@@ -3,12 +3,18 @@ package Packets;
 
 
 public class HomePacket extends Packets
-{	
+{
+	private static final long serialVersionUID = 4367432052684494554L;
+
+	public NewsPacket[] news;
+	public LadderPacket ladder;
+	public ServerInfoPacket serverInfos;
+	
 	public HomePacket(String ip) 
 	{
 		super(ip);
 		
-		ladderPlayers = new PlayerPacket[9];
+		ladder = new LadderPacket();
 	}
 
 	public HomePacket() 
@@ -16,21 +22,12 @@ public class HomePacket extends Packets
 		super();
 	}
 	
-	/**
-	 * 
-	 */
-	private static final long serialVersionUID = 4367432052684494554L;
-
-	public NewsPacket[] news;
-	public PlayerPacket[] ladderPlayers;
-	public ServerInfoPacket serverInfos;
-	
 	@Override
 	public String toString()
 	{
 		return super.toString() 
 				+ "\n"
-				+ "- [ladderPlayers]: " + ladderPlayers.toString()
+				+ "- [ladderPlayers]: " + ladder.toString()
 				+ "\n";
 	}
 }
